Move ErrorComponent to separate file
