+++ src/containers/ArticlePage/TopicArticlePage/components/TopicArticleTaxonomy.tsx
@@ -18,6 +18,7 @@ import {
   addTopicToTopic,
   addSubjectTopic,
   addTopic,
+  queryTopics,
 } from '../../../../modules/taxonomy';
 import {
   sortByName,
@@ -115,6 +116,7 @@ class TopicArticleTaxonomy extends Component<Props, State> {
 
   fetchTaxonomy = async () => {
     const {
+      articleId,
       article: { language },
     } = this.props;
     if (!language) return;
@@ -124,7 +126,7 @@ class TopicArticleTaxonomy extends Component<Props, State> {
         fetchSubjects(language),
       ]);
 
-      const topics = this.props.article.taxonomy?.topics ?? [];
+      // TODO: const topics = this.props.article.taxonomy?.topics ?? [];
 
       const sortedSubjects = subjects.filter(subject => subject.name).sort(sortByName);
       const activeTopics = topics.filter(t => t.path);
