+++ src/containers/StructurePage/StructureContainer.tsx
@@ -11,15 +11,13 @@ import React, { useContext, useState, useRef } from 'react';
 import { useTranslation } from 'react-i18next';
 import { OneColumn } from '@ndla/ui';
 import { RouteComponentProps, withRouter } from 'react-router-dom';
-import { AlertCircle, Taxonomy } from '@ndla/icons/editor';
+import { Taxonomy } from '@ndla/icons/editor';
 import { Structure } from '@ndla/editor';
-import Tooltip from '@ndla/tooltip';
 import { Switch } from '@ndla/switch';
-import { colors, spacing } from '@ndla/core';
+import { colors } from '@ndla/core';
 import { useEffect } from 'react';
 import { useLayoutEffect } from 'react';
 import { DropResult } from 'react-beautiful-dnd';
-import styled from '@emotion/styled';
 import handleError from '../../util/handleError';
 import StructureResources from './resourceComponents/StructureResources';
 import FolderItem from './folderComponents/FolderItem';
@@ -35,7 +33,7 @@ import {
   updateSubjectTopic,
 } from '../../modules/taxonomy';
 import { groupTopics, getCurrentTopic, getSubtopics } from '../../util/taxonomyHelpers';
-import { fetchDraft, fetchUserData, updateUserData } from '../../modules/draft/draftApi';
+import { fetchUserData, updateUserData } from '../../modules/draft/draftApi';
 import { REMEMBER_FAVOURITE_SUBJECTS, TAXONOMY_ADMIN_SCOPE } from '../../constants';
 import Footer from '../App/components/Footer';
 import { ButtonAppearance } from '../../components/Accordion/types';
@@ -46,50 +44,7 @@ import {
   TaxonomyMetadata,
 } from '../../modules/taxonomy/taxonomyApiInterfaces';
 import { LocaleContext, UserAccessContext } from '../App/App';
-
-const StyledWarnIcon = styled(AlertCircle)`
-  height: ${spacing.nsmall};
-  width: ${spacing.nsmall};
-  fill: ${colors.support.red};
-`;
-
-const ErrorComponent = ({ contentUri, isSubject }: { contentUri?: string; isSubject: boolean }) => {
-  const { t } = useTranslation();
-  const wrongTooltip = t('taxonomy.info.wrongArticleType', {
-    placedAs: t(`articleType.topic-article`),
-    isType: t(`articleType.standard`),
-  });
-  const [error, setError] = useState(false);
-
-  useEffect(() => {
-    let shouldUpdateState = true;
-
-    const fetchAndSetError = async (contentUri: string) => {
-      const articleId = contentUri.split(':').pop();
-      if (articleId) {
-        const fetched = await fetchDraft(Number(articleId));
-        if (fetched.articleType !== 'topic-article') {
-          if (shouldUpdateState) setError(true);
-        }
-      }
-    };
-
-    if (isSubject || !contentUri) return;
-
-    fetchAndSetError(contentUri);
-    return () => {
-      shouldUpdateState = false;
-    };
-  }, [isSubject, contentUri]);
-
-  if (!error) return null;
-
-  return (
-    <Tooltip tooltip={wrongTooltip}>
-      <StyledWarnIcon title={wrongTooltip} />
-    </Tooltip>
-  );
-};
+import StructureErrorIcon from './folderComponents/StructureErrorIcon';
 
 interface Props extends RouteComponentProps<StructureRouteParams> {}
 
@@ -324,7 +279,7 @@ export const StructureContainer = ({ match, location, history }: Props) => {
               highlightMainActive
               toggleFavorite={toggleFavorite}
               favoriteSubjectIds={favoriteSubjects}
-              renderBeforeTitles={isTaxonomyAdmin ? ErrorComponent : undefined}
+              renderBeforeTitles={isTaxonomyAdmin ? StructureErrorIcon : undefined}
               renderListItems={({
                 pathToString,
                 parent,
